Collate different model objects into a single .csv file

The function looks in the output folder and collects all model objects given by the pattern and collates them into a single .csv file for easy comparison. To make comparison easier, the function will create a vector of unique parameter names and match parameters from all models to the vector. This ensures that the same parameters from different models are in the same row in the output file.

collate(path = NULL, pattern = NULL, digits = 4)

Arguments

path
The relative path (from working directory) to the folder where the outputs are stored. Default is NULL, which means outputs are found in the root working directory

pattern
A regular expression. Only file names which match the regular expression will be returned.

digits
An integer giving the number of digits to round results