Inspect a list

A useful function for exploratory work or when coding. Takes a list as an argument and prints a glimpse of the data contained in each list element and the dimensions of each list element. If the list contains more than 5 list elements, only the first 5 elements are printed to console with no warning. The function will print a maximum of 10 values if the list element is a vector and 10 rows and 5 columns if the list element is a matrix.

inspect_list(x)

Arguments

x
A list of vectors or matrices

Details

NOTE: There are currently no checks built in to handle a list of lists. In this case, everything in the sublists will be printed with no checks or warnings.

Examples

x <- list(
  matrix(runif(100), nrow = 20, ncol = 5),
  matrix(runif(100), nrow = 10, ncol = 10)
)

inspect_list(x)

#> [[1]]
#> [[1]][[1]]
#>             [,1]      [,2]      [,3]     [,4]      [,5]
#>  [1,] 0.08075014 0.2898923 0.6801629 0.825199 0.5279170
#>  [2,] 0.83433304 0.6783804 0.4988456 0.273818 0.6006375
#>  [3,] 0.60076089 0.7353196 0.6416793 0.570045 0.2613714
#>  [4,] 0.15720844 0.1959567 0.6602843 0.335719 0.2900502
#>  [5,] 0.00739944 0.9805397 0.0960242 0.596263 0.4800752
#>  [6,] 0.46639350 0.7415215 0.7656002 0.191518 0.9200055
#>  [7,] 0.49777739 0.0514463 0.7696748 0.947764 0.4007202
#>  [8,] 0.28976724 0.5302125 0.9907123 0.542480 0.2131727
#>  [9,] 0.73288199 0.6958239 0.9705209 0.544603 0.6717668
#> [10,] 0.77252151 0.6885560 0.3891828 0.278597 0.0586141
#> 
#> [[1]][[2]]
#> [1] 20  5
#> 
#> 
#> [[2]]
#> [[2]][[1]]
#>           [,1]      [,2]     [,3]      [,4]      [,5]
#>  [1,] 0.805680 0.7087812 0.858666 0.1645692 0.0706812
#>  [2,] 0.814051 0.8742059 0.566894 0.6632066 0.9968915
#>  [3,] 0.403911 0.0114795 0.252997 0.8565750 0.6118524
#>  [4,] 0.218431 0.8882496 0.918803 0.9265464 0.1725588
#>  [5,] 0.418361 0.9963469 0.867350 0.5523776 0.9094410
#>  [6,] 0.668871 0.5001915 0.248539 0.5770657 0.0374512
#>  [7,] 0.507650 0.3589670 0.402881 0.6874477 0.5935538
#>  [8,] 0.660359 0.7749130 0.769630 0.2447182 0.2369776
#>  [9,] 0.511791 0.5844753 0.119485 0.0446172 0.9062973
#> [10,] 0.835552 0.6339764 0.194695 0.9098546 0.8188730
#> 
#> [[2]][[2]]
#> [1] 10 10
#> 
#>