Ordered logit

This is a function to calculate an ordered logit model.

ordered_logit(ordered_outcome, x, tau, lvls = NULL)

Arguments

ordered_outcome
The ordered outcome variable

x
A vector used to shift the the estimated thresholds. The default value is 0, which implies no shifting.

tau
A vector of parameters. The length of the vector is equal to the number of levels of x - 1. No checks beyond 'non-conformable arrays' is written to ensure this is the case.

lvls
A vector giving the levels of the ordered outcome. If lvls is missing (default), then the code assumes that they are ascending integers.

Details

The function is based on the core of the 'apollo_ol' from the Apollo package.