Prepares the draws

In a model using simulation, we need to prepare a set of random draws to use when simulating/approximating the distributions/integrals. The function is a convenient wrapper around the function make_draws.

prepare_draws(n_id, n_ct, n_draws, type, rpar, seed)

Arguments

n_id

An integer giving the nubmer of respondents

n_ct

An integer giving the number of choice situations

n_draws

An integer giving the number of draws per individual

type

A character string giving the type of draws to use

rpar

A named list giving the type of distribution

seed

An integer giving the seed to change the scrambling of the sobol sequence

Value

A list of matrixes matching 'rpar' where each matrix has the dimensions n_id*n_ct x n_draws

Details

Depending on the model options, the function will transform the supplied uniform draws to either standard uniform, standard normal or standard triangular. It is up to the user to specify the correct distributions in the log-likelihood function.