Search for starting values

The function searches for new starting values. Either in a random fashion or based on a smarter search algorithm. The smart search algorithm is a rewritten version of the search algorithm implemented in the Apollo package. The function is rewritten to work with the current modeling framework.

search_start_values(
  ll,
  estim_env,
  model_options,
  control = NULL,
  type = "simple",
  n_candidates = 100,
  n_return = 10,
  multiplier = 1
)

Arguments

ll

This is the 'raw' log-likelihood function passed to the estimation routine. It is important that the user takes into account whether the optimization routine is a minimizer (e.g. 'ucminf') or a maximizer (e.g. 'maxlik').

estim_env

An estimation environment or list of estimation environments returned from prepare

model_options

A list of model options. Note, this list is validated for a second time here to set some necessary defaults. See validate for details.

control

A list of control options that are passed to set_controls.

type

A character string indicating whether to use a simple or a smart search algorithm. Default is 'simple'

n_candidates

An integer giving the number of candidates to evaluate. Default is 100.

n_return

An integer giving the number of parameter vectors to return. The default is 10

multiplier

A double indicating a multiplier for the 'simple' search algorithm. The default is 1.

Value

A matrix of starting values

References

Hess, S. & Palma, D., 2019, Apollo: A flexible, powerful and customisable f reeware package for choice model estimation and application, Journal of Choice Modelling, 32 Bierlaire, M., Thémans, M. & Zufferey, N., 2010, A heuristic for nonlinear global optimization, INFORMS Journal on Computing, 22(1):